+++ daedalus-server/src/test/java/com/daedalus/server/WebUiPaintPinTest.java
@@ -436,7 +436,9 @@ class WebUiPaintPinTest {
                     .contains("world.drive")
                     .contains("row(box, \"drive\"")
                     .contains("world.driveActor")
-                    .contains("row(box, \"actor\"");
+                    .contains("row(box, \"actor\"")
+                    .contains("world.driveAt")
+                    .contains("row(box, \"at\"");
             assertThat(world).doesNotContain("/world/\" + WORLD + \"/parcels\"");
             assertThat(html).contains("#worldBox .place { color: #94612e; -webkit-user-drag: none")
                     .contains("#worldBox .slab { color: #94612e; -webkit-user-drag: none");